move IndexFilter to the index module
