+++ fast_forward/src/index/mod.rs
@@ -37,6 +37,48 @@ use std::{
     slice,
 };
 
+/// `IndexFilter` means, that you get an `Iterator` over all `Items` which exists for a given list of indices.
+pub trait IndexFilter: Index<usize, Output = Self::Item> {
+    type Item;
+
+    /// Returns a `Iterator` over all `Items` with the given index list.
+    fn filter<'i>(&'i self, indices: SelectedIndices<'i>) -> Filter<'i, Self>
+    where
+        Self: Sized,
+    {
+        Filter::new(self, indices)
+    }
+}
+
+pub struct Filter<'i, F> {
+    pos: usize,
+    list: &'i F,
+    indices: SelectedIndices<'i>,
+}
+
+impl<'i, F> Filter<'i, F> {
+    pub const fn new(list: &'i F, indices: SelectedIndices<'i>) -> Self {
+        Self {
+            pos: 0,
+            list,
+            indices,
+        }
+    }
+}
+
+impl<'i, F> Iterator for Filter<'i, F>
+where
+    F: IndexFilter,
+{
+    type Item = &'i F::Item;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        let idx = self.indices.get(self.pos)?;
+        self.pos += 1;
+        Some(&self.list[*idx])
+    }
+}
+
 #[derive(Debug, Default, Clone)]
 #[repr(transparent)]
 pub struct SelectedIndices<'i>(Cow<'i, [usize]>);
